requirements/local.txt
-r base.txt

Werkzeug[watchdog]==2.2.2
ipdb==0.13.9
psycopg2==2.9.5

# Testing
# ------------------------------------------------------------------------------
mypy==0.982
django-stubs==1.12.0
pytest==7.2.0
pytest-sugar==0.9.5
djangorestframework-stubs==1.7.0



# Code quality
# ------------------------------------------------------------------------------
flake8==5.0.4
coverage==6.5.0


# Django
# ------------------------------------------------------------------------------
factory-boy==3.2.1
django-debug-toolbar==3.7.0
django-extensions==3.2.1
django-coverage-plugin==2.0.3
pytest-django==4.5.2
